Also ignore form validation pseudo classes
`dePseudify()` strips several pseudo classes off of selectors such that
the processing done by `filterUnusedSelectors()` can determine whether
the "parent" selector is actually used or not.

This expands the list of stripped pseudo classes to include `:required`,
`:invalid`, and `:valid`, which can be used in conjunction with form
validation states.

Without this, the `document.querySelector()` call in `findAll()` raises
an exception, erroneously causing the selector to be considered as used.